move the storage specific query processors to their own files
